refactor: 소켓, 로그인 관련 유틸함수 디렉토리명 변경, 코드스타일 통일
