+++ routes/controllers/loginController.js
@@ -2,7 +2,7 @@ const jwt = require("jsonwebtoken");
 const passport = require("passport");
 
 const ERROR = require("../../constants/error");
-const validationCheck = require("../../utils/validationCheck");
+const validationCheck = require("../../services/validationCheck");
 
 module.exports = {
   login: async function (req, res, next) {
